backend/requirements.txt
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.34.0,<1.0
pydantic>=2.12.0,<3.0
pydantic-settings>=2.8.0,<3.0
python-dotenv>=1.0.1,<2.0
PyYAML>=6.0.2,<7.0
pytest>=8.3.0,<9.0
httpx>=0.28.0,<1.0
